[FEAT] : friend
